requirements.txt
aniso8601==9.0.1
asn1crypto==1.4.0
attrs==21.2.0
click==8.0.1
Flask==2.0.1
Flask-Cors==3.0.10
flask-restplus==0.13.0
itsdangerous==2.0.1
Jinja2==3.0.1
jsonschema==3.2.0
MarkupSafe==2.0.1
pg8000==1.21.0
psycopg2-binary==2.9.1
pyrsistent==0.18.0
pytz==2021.1
scramp==1.4.0
six==1.16.0
Werkzeug==2.0.1
